requirements.txt
numpy==1.26.4
scikit-surprise
flask
unicorn
pandas
scikit-learn
fastapi
uvicorn
